% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mod.R
\name{add_fit}
\alias{add_fit}
\title{Estimates and Add F-values from a Linear Model to an og_set class element}
\usage{
add_fit(ogset, log_scale = FALSE)
}
\arguments{
\item{ogset}{a \code{ogset} class element}

\item{log_scale}{logical, should the data be natural log trasnsformed before fitting the model? (actually \code{log(data + 1)} transformed)

The info on the \code{log_scaled} parameters get stored in the metadata slot}
}
\description{
Estimates and Add F-values from a Linear Model to an og_set class element
}
