% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{collapse_orthologs}
\alias{collapse_orthologs}
\title{Sum Expression of Orthologs in an ogset class elements}
\usage{
collapse_orthologs(og_set, mc.cores = 1)
}
\arguments{
\item{og_set}{a \code{ogset} class element}

\item{mc.cores}{\code{numeric}, the number of threads}
}
\description{
#' \code{collapse_orthologs} takes ogset class element as input
It sums the expression of all the genes in one orthogroup across the samples
and outputs an ogset class element containing the orthologs expression matrix
}
\details{
#' Some step are considerably slow and are parallelized with the function
\code{mclapply} from the package \code{parallel} the number of threads defaults to 1.
}
