% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{collapse_orthologs_3}
\alias{collapse_orthologs_3}
\title{Sum Expression of Orthologs and Make one Expression Matrix for two species}
\usage{
collapse_orthologs_3(eset1, eset2, ogroups, n_cols = 18, mc.cores = 1)
}
\arguments{
\item{eset1}{a \code{data.frame} with the expression matrix for species 1}

\item{eset2}{a \code{data.frame} with the expression matrix for species 2}

\item{ogroups}{a \code{list} of orthogroups, the gene names in the list must match the \code{row.names} of the datasets}

\item{n_cols}{numeric, the number of sample for each species}

\item{mc.cores}{numeric, the number of threads}
}
\description{
\code{collapse_orthologs_3} takes two expression matrices and a list of orthogroups.
It sums the expression of the genes in the orthogroups for the two species separately
and outputs a single dataset with orthogroup-wise expression
}
\details{
The function assumes that the info on the species are contained in the first two
characters of the gene names, and that the gene names are the rownames of the datasets.
The samples for each dataset should be the same The numbers of samples per dataset is the
input n_cols. Some step are considerably slow and are parallelized with the function
\code{mclapply} from the package \code{parallel} the number of threads defaults to 1
}
