% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{cross_validation: Implement k-fold cross validation, with specified k, returning the scores
for each fold.}
\usage{
cross_validation(X, y, k = 3, shuffle = TRUE, random_state = 0)
}
\arguments{
\item{X:}{features data frame}

\item{y:}{target data frame}

\item{k:}{number of splits}

\item{shuffle:}{boolean}

\item{random_state:}{integer}
}
\value{
vector of k scores
}
\description{
cross_validation: Implement k-fold cross validation, with specified k, returning the scores
for each fold.
}
\examples{
data <- gen_data(100)
X <- data.frame(data[[1]])
y_vec <- data[[2]]
y <- data.frame(y = y_vec)
cross_validation(X = X, y = y, k = 5)
}
