% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossR.R
\docType{package}
\name{CrossR}
\alias{CrossR}
\alias{CrossR-package}
\title{\code{CrossR} package\if{html}{\figure{logo.png}{options: width="25\%" align="right" alt="Figure: logo.png"}}}
\description{
HIIT Exercises for the R User
}
\details{
Instructors can leverage CrossR to lead R users through high-intensity interval
training (HIIT) exercise sessions. These sessions guide R users through simple,
hands-on coding activities repetitiously for 20-90 seconds at a time. The goal
is to help R users become more fluent with the R programming language and
RStudio's shortcuts, while keeping things challenging and fun.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeffboichuk/CrossR}
  \item Report bugs at \url{https://github.com/jeffboichuk/CrossR/issues}
}

}
\author{
\strong{Maintainer}: Jeff Boichuk \email{boichuk@virginia.edu}

Authors:
\itemize{
  \item Steven M. Mortimer \email{reportmort@gmail.com}
}

Other contributors:
\itemize{
  \item Research for Sustainable Commerce at UVA \email{R4SC@virginia.edu} [copyright holder, funder]
}

}
