% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{explore_ogroups}
\alias{explore_ogroups}
\title{Plot an Histogram of Orthogroup Dimension}
\usage{
explore_ogroups(groups, main = "dimension of orthogroups")
}
\arguments{
\item{groups}{a \code{list} of orthogroups}

\item{main}{an overall title for the plot}
}
\description{
\code{explore_ogroups} is a wrapper for the \code{plot} method for \code{table}
}
