% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{gen_data}
\alias{gen_data}
\title{gen_data(): returns data X, y for testing.}
\usage{
gen_data(N, perfect = FALSE)
}
\arguments{
\item{N}{number of obervations}

\item{perfect}{get perfect linear data or not}
}
\value{
a list consisting of X and y (X - a dataframe, y - a numeric vector)
}
\description{
gen_data(): returns data X, y for testing.
}
\examples{
data = gen_data(100)
X <- data[[1]]
y <- data[[2]]

}
