% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_df}
\alias{get_df}
\title{Returns a Tidy Dataset in Long Format with the Expression, Statistics and
Metadata for Selected Orthogoups}
\usage{
get_df(ogset, orthogroups, rank_stat = NULL, use_annos = FALSE)
}
\arguments{
\item{ogset}{The \code{ogset} class element that contains the expression data.}

\item{orthogroups}{The ortogroups for which we want to extract expression data}

\item{rank_stat}{The ranking statistics that we want to extract, defaults to \code{NULL}.}

\item{use_annos}{Should functional annotations be included in the dataset? Defaults to \code{FALSE}.}
}
\description{
The long dataset format provided by \code{get_df} can be used straight away to plot orthogroup
expression with \code{ggplot2}.
}
