% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-encouragement.R
\name{get_encouragement}
\alias{get_encouragement}
\title{Get Encouragement}
\usage{
get_encouragement()
}
\value{
A character string message of encouragment
}
\description{
This function returns one of many different messages that give the user positive
feedback letting them know that they are doing great and should keep going!
}
\examples{
get_encouragement()
}
