% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_gene_df}
\alias{get_gene_df}
\title{For all Genes in One Orthogroup, Returns a Tidy Dataset in Long Format with the Expression,
Statistics and Metadata}
\usage{
get_gene_df(ogset, orthogroup)
}
\arguments{
\item{ogset}{The \code{ogset} class element that contains the expression data.}

\item{orthogroup}{The ortogroups for which we want to extract expression data}
}
\description{
NOTE the function returns NULL, still writing it.
}
