% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{get_nrows}
\alias{get_nrows}
\title{get_nrows(): returns the number of rows of a dataframe or the length of an atomic vector.}
\usage{
get_nrows(data)
}
\arguments{
\item{data}{a dataframe or an atomic vector,}
}
\value{
number of observations
}
\description{
get_nrows(): returns the number of rows of a dataframe or the length of an atomic vector.
}
\examples{
nrows = get_nrows(1:10)
nrows = get_nrows(mtcars)

}
