% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_all_stages}
\alias{ggplot_all_stages}
\title{Plot the Expression of One Orthogroup in Both Species with ggplot2}
\usage{
ggplot_all_stages(orthogroups, ogset, species_var, condition_var,
  colours = c("darkblue", "#56B4E9"), use_annos = FALSE)
}
\arguments{
\item{orthogroups}{a \code{character} vector with the orthogroups for the plot}

\item{ogset}{an ogset class element}

\item{species_var}{a \code{character} string with the name of column of coldata that encodes
for the species from which the samples originates.}

\item{condition_var}{a \code{character} string with the name of column of coldata that encodes the experimental
conditions under which you are comparing the samples}

\item{colours}{a \code{character} vector with the colours for the plot}

\item{use_annos}{\code{logical}; if \code{TRUE} the functional annotation from og_annos will
be displayed in the plot subtitle. Defaults to \code{FALSE}}
}
\description{
\code{ggplot_all_stages} is a wrapper for ggplot2 that plots a stripchart
and a line plot of the expression of an orthogroup in both species.
}
\details{
This function takes one (ot many) orthogroup IDs and an ogset class element as arguments.
It plots a stripchart of the expression of all the stages.

You can also provided a vector of multiple orthogroup IDs, \code{ggplot_all_stages} will plot
all of them using faceting.

The function returns a ggplot object that can be further customized with all basic
ggplot2 functionalities.
}
