% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_ogroup_genes}
\alias{ggplot_ogroup_genes}
\title{Plot the Expression of all Genes within one Orthogroup with ggplot2}
\usage{
ggplot_ogroup_genes(orthogroup, ogset)
}
\arguments{
\item{orthogroup}{the orthogroup that you want to plot}

\item{ogset}{the \code{ogset} class element that contains the expression data}
}
\description{
\code{ggplot_ogroup_genes} is a wrapper for \code{ggplot2} that plots a stripchart
of the expression of all the gene contained in a orthogroup
}
