% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mod.R
\name{make_fit}
\alias{make_fit}
\title{Calculate F-values from a Linear Model on an Orthogroup Expression Set}
\usage{
make_fit(dat, coldata, design, mc.cores = 1)
}
\arguments{
\item{dat}{\code{data.frame} contains data for any gene}

\item{coldata}{\code{data.frame} categorical variables with info for every samples}

\item{design}{\code{formula}, encodes the assumption on how the variables in coldata
explain the observed gene expression}

\item{mc.cores}{the number of cores to be used in \code{parallel::mclapply}, defaults to 1}
}
\description{
Calculate F-values from a Linear Model on an Orthogroup Expression Set
}
