% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{make_TPM_df}
\alias{make_TPM_df}
\title{Make Expression Matrix from All Salmon Output Folders}
\usage{
make_TPM_df(path)
}
\arguments{
\item{path}{character string, the path for the folder that contains the Salmon results
for all the samples}
}
\description{
\code{Make_TPM_df} is a wrapper for \code{\link{get_TPM}}, it takes as input the path
for the folder that contain the Salmon output folders for all your sample and returns
the TPM expression matrix as a \code{data.frame} object
}
