% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_orthogroups}
\alias{parse_orthogroups}
\title{Parse and Load an Orthogroup CSV File Provided by Orthofinder}
\usage{
parse_orthogroups(path_2_ogroups)
}
\arguments{
\item{path_2_ogroups}{a character string that provides the path to the orthogroup
file}
}
\description{
\code{parse_orthogroups} parses an orthogroup \strong{CSV} file provided by Orthofinder and
outputs the result as a list of orthogroups
}
\section{Warning}{
 Orthofinder outputs both and OrthologousGroup.txt file and
OrthologousGroup.csv file, this function is designed to parse the \strong{OrthologousGroup.csv} file
}

\section{Note}{
 \url{https://github.com/davidemms/OrthoFinder}; Emms, D.M. and Kelly, S. (2015)
OrthoFinder: solving fundamental biases in whole genome comparisons dramatically improves orthogroup
inference accuracy, Genome Biology 16:157
}

