% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_keyword}
\alias{plot_keyword}
\title{Plot the Expression of Every Orthogroup that Contains a Keyword in their Annos}
\usage{
plot_keyword(keyword, dset, clusters)
}
\arguments{
\item{keyword}{a character string}

\item{dset}{a numeric data.frame with ogroups as rownames}

\item{clusters}{a named numeric vector}
}
\description{
\code{plot_keyword} is a wrapper for \code{\link{plot_all_stages}} that applies this
function to all the orthogroups in the dataset that contain a user specified keyword
in their functional annotation
}
\details{
The function uses grep to search for the \code{keyword} in the functional annotation
of the orthogroups
}
