% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{px_2_tx}
\alias{px_2_tx}
\title{Converts Protein ID into Transcript ID in Orthogroup List}
\usage{
px_2_tx(ogroups, anno_file, tx_id, px_id, mc.cores = 1)
}
\arguments{
\item{anno_file}{a character string containing the path to the annotation file in tabulated format}

\item{tx_id}{a \code{character} string indicating the name of the column of transcript ids in the annotation file}

\item{px_id}{a \code{character} string indicating the name of the column of protein ids in the annotation file}

\item{mc.cores}{\code{numeric}, the number of threads used for \code{mclapply}}

\item{groups}{a \code{list} of orthogroups}
}
\description{
Orthogroups are generated on protein sequence, \code{px_2_tx} converts protein ID to transcript ID in an orthogroup list
}
