% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortcuts}
\alias{shortcuts}
\title{Shortcuts Table}
\format{a \code{tbl_df} with 140 rows and 4 variables}
\usage{
data(shortcuts)
}
\description{
This data.frame contains the RStudio keyboard shortcuts for both Mac OSX and Windows.
}
\details{
\itemize{
  \item category: one of ten categorizations of shortcuts based on behavior
  \item description: a brief description of what the shortcut does
  \item windows: the shortcut keys on Windows
  \item mac: the shortcut keys on Mac
}
}
\examples{
\dontrun{
data(shortcuts)

head(shortcuts, 3)
#   category                      description windows           mac
# 1  Console           Move cursor to Console  Ctrl+2        Ctrl+2
# 2  Console                    Clear console  Ctrl+L        Ctrl+L
# 3  Console Move cursor to beginning of line    Home  Command+Left
}
}
\keyword{datasets}
