% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{sum_groups}
\alias{sum_groups}
\title{Sum Expression of Orthologs}
\usage{
sum_groups(eset, ogroups, mc.cores = 1)
}
\arguments{
\item{eset}{a \code{data.frame} with the a gene-wise  or transcript-wise expression matrix}

\item{ogroups}{a \code{list} of orthogroups, the gene names in the list should match the \code{row.names} of the datasets}

\item{mc.cores}{\code{numeric}, the number of threads}
}
\description{
\code{sum_groups} takes an expression matrix and a list of orthogroups.
It sums the expression of all the genes in one orthogroup across the samples
and outputs a orthogroup-wise expression matrix
}
\details{
Some step are considerably slow and are parallelized with the function
\code{mclapply} from the package \code{parallel} the number of threads defaults to 1.
The gene/transcripts IDs in the orthogroups are searched for a match in the \code{rownames}
of the expressiom matrix using \code{\link{grep}}
}
