% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cv.R
\name{summary_cv}
\alias{summary_cv}
\title{summary_cv: Summary statistics of cross-validation scores}
\usage{
summary_cv(scores)
}
\arguments{
\item{scores}{vector of cross-validation scores}
}
\value{
list of summary statistics consisting of mean, median, and standard deviation
}
\description{
summary_cv: Summary statistics of cross-validation scores
}
\examples{
cv_scores = c(0.97, 0.96, 0.98)
summary_cv(scores = cv_scores)
}
