% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{train_test_split: Split X and ydata into two portions according to input ratio.
Default is for the split to include shuffling.}
\usage{
train_test_split(X, y, test_size = 0.25, random_state = 0, shuffle = TRUE)
}
\arguments{
\item{X}{features, a dataframe or a vector,}

\item{y}{target, a vector or a dataframe.}

\item{test_size}{float between 0 and 1}

\item{random_state}{A integer for seting the random seed.}

\item{shuffle}{boolean, when TRUE, shuffle the data.}
}
\value{
list of split data in the order X_train, X_test, y_train, y_test.
}
\description{
train_test_split: Split X and ydata into two portions according to input ratio.
Default is for the split to include shuffling.
}
\examples{
X = data.frame(X = rnorm(100, 0, 10))
y =data.frame(y = 2 * X$X + rnorm(100))
split_data = train_test_split(X, y, test_size = 0.5)

}
