% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mod.R
\name{add_fit}
\alias{add_fit}
\title{Add F-value from Linear Model to Orthologroup Expression Set}
\usage{
add_fit(dat, coldata)
}
\description{
Add F-value from Linear Model to Orthologroup Expression Set
}
