% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{check_ogset}
\alias{check_ogset}
\title{Check Validity of an \code{ogset} class element}
\usage{
check_ogset(object)
}
\arguments{
\item{object}{an \code{ogset} class element}
}
\description{
Check Validity of an \code{ogset} class element
}
