% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-shortcut.R
\name{find_shortcut}
\alias{find_shortcut}
\title{Find Keyboard Shortcut}
\usage{
find_shortcut(keyword, os = get_os())
}
\arguments{
\item{keyword}{character; a case-insensitive regex pattern identifying rows
to filter}

\item{os}{character; a string vector specifying \code{"windows"},
\code{"mac"}, or both to return shortcuts that work on the user's operating
system or the specified operating system(s)}
}
\value{
A \code{tbl_df} of all shortcuts matching the keyword and operating
  system(s)
}
\description{
This function returns all keyboard shortcuts that match the keyword and
operating system(s) specified.
}
\examples{
find_shortcut(keyword = "clear", os = "mac")
}
