% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{get_ncols}
\alias{get_ncols}
\title{get_ncols(): returns the number of columns of data.}
\usage{
get_ncols(data)
}
\arguments{
\item{data}{a dataframe or an atomic vector,}
}
\value{
number of observations
}
\description{
get_ncols(): returns the number of columns of data.
}
\examples{
ncols = get_ncols(1:10)
ncols = get_ncols(mtcars)

}
