% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_TPM}
\alias{get_TPM}
\title{Extracts TPM for a Salmon Output File}
\usage{
get_TPM(path)
}
\arguments{
\item{path}{a character string containing the path to the SALMON output file}
}
\description{
\code{get_TPM} is a wrapper for \code{\link{read.csv}} that loads and extracts
TPM from a salmon output file provided in \code{path}
}
