% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_orthogroups}
\alias{parse_orthogroups}
\title{Parse and Load an Orthogroup File Provided by Orthofinder}
\usage{
parse_orthogroups(path_2_ogroups)
}
\arguments{
\item{path_2_ogroups}{a character string that provides the path to the orthogroup
file}
}
\description{
\code{parse_orthogroups} parses an orthogroup file provided by Orthofinder and
outputs the result as a list of orthogroups
}
