% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all_stages}
\alias{plot_all_stages}
\title{Plot the Expression of One Orthogroup in Both Species}
\usage{
plot_all_stages(ogroup, dset, ylab = "normalized expression",
  clusters = "cls_here")
}
\arguments{
\item{ogroup}{a character string}

\item{dset}{a numeric data.frame with ogroups as rownames}

\item{ylab}{a character string}

\item{clusters}{a named numeric vector}
}
\description{
\code{plot_all_stages} is a wrapper for \code{stripchart} and plots a stripchart
and a line plot of the expression of an orthogroup across all samples
}
\details{
This function takes an orthogroup ID and the expression matrix.
It plots a stripchart of the expression of all the stages; The title of the
stripchart contains details on: how many genes are in the orthogroup, the cluster
that contain the ogroup (if any), the F-value for the interaction term.
}
