% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_og_genes}
\alias{plot_og_genes}
\title{plot the expression of all genes within one orthogroup}
\usage{
plot_og_genes(ogroup)
}
\arguments{
\item{ogroup}{a character string}
}
\description{
\code{plot_og_gene} is a wrapper for \code{stripchart} and plots a stripchart
of the expression of all the gene within a orthogroup for each species in which
expression data are available
}
