% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_og_genes}
\alias{plot_og_genes}
\title{Plot the Expression of all Genes within one Orthogroup}
\usage{
plot_og_genes(ogroup, ogset, ylab = "TPM")
}
\arguments{
\item{ogroup}{a character string with the ID of the orthogroups that contains
the genes to be plotted}

\item{ogset}{an ogset class element}

\item{ylab}{a character string, the name of the y-axis in the plots}
}
\description{
\code{plot_og_gene} is a wrapper for \code{stripchart} and plots a stripchart
of the expression of all the gene within a orthogroup for each species in which
expression data are available.
}
